+++ CHANGELOG.md
@@ -100,3 +100,5 @@
 ## 0.4.1
 
 - The time widget is now editable when replay is paused
+- Automatically pause/unpause replay only when in connected state
+